src/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.8.1
cffi==1.16.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.0
Flask-WTF==1.2.1
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.22
Werkzeug==3.0.2
WTForms==3.1.2
zipp==3.18.1
